ajuste no usuariorepository

+++ users-vem/src/main/java/com/urbanape/usersvem/repository/UsuarioRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.urbanape.usersvem.model.UsuarioModel;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioModel, Long> {
+public interface UsuarioRepository extends JpaRepository<UsuarioModel, Integer> {
 }
